# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/users": {
      "get": {
        "description": "Retrieve all registered users. Admin access only.",
        "operationId": "UsersController_findAll_api/v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Users retrieved successfully."
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Bad Request",
                  "status": 400,
                  "detail": "Validation failed",
                  "instance": "/api/v1/tasks",
                  "errors": [
                    "email must be an email",
                    "password must be longer than or equal to 8 characters"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden. Admin role required."
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Not Found",
                  "status": 404,
                  "detail": "Resource not found",
                  "instance": "/api/v1/tasks/..."
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Conflict",
                  "status": 409,
                  "detail": "The request conflicts with the current state of the resource",
                  "instance": "/api/v1/tasks"
                }
              }
            }
          },
          "412": {
            "description": "Precondition Failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Precondition Failed",
                  "status": 412,
                  "detail": "Task has been modified. Please refresh and try again.",
                  "instance": "/api/v1/tasks/..."
                }
              }
            }
          },
          "428": {
            "description": "Precondition Required",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Precondition Required",
                  "status": 428,
                  "detail": "If-Match header is required",
                  "instance": "/api/v1/tasks/..."
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Too Many Requests",
                  "status": 429,
                  "detail": "Too many requests",
                  "instance": "/api/v1/tasks"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Internal Server Error",
                  "status": 500,
                  "detail": "Internal server error",
                  "instance": "/api/v1/tasks"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get all users",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/{id}": {
      "delete": {
        "description": "Delete a user by ID. Admin access only.",
        "operationId": "UsersController_deleteUser_api/v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted successfully."
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Bad Request",
                  "status": 400,
                  "detail": "Validation failed",
                  "instance": "/api/v1/tasks",
                  "errors": [
                    "email must be an email",
                    "password must be longer than or equal to 8 characters"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden. Admin role required."
          },
          "404": {
            "description": "User not found."
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Conflict",
                  "status": 409,
                  "detail": "The request conflicts with the current state of the resource",
                  "instance": "/api/v1/tasks"
                }
              }
            }
          },
          "412": {
            "description": "Precondition Failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Precondition Failed",
                  "status": 412,
                  "detail": "Task has been modified. Please refresh and try again.",
                  "instance": "/api/v1/tasks/..."
                }
              }
            }
          },
          "428": {
            "description": "Precondition Required",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Precondition Required",
                  "status": 428,
                  "detail": "If-Match header is required",
                  "instance": "/api/v1/tasks/..."
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Too Many Requests",
                  "status": 429,
                  "detail": "Too many requests",
                  "instance": "/api/v1/tasks"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Internal Server Error",
                  "status": 500,
                  "detail": "Internal server error",
                  "instance": "/api/v1/tasks"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Delete a user",
        "tags": [
          "Users"
        ]
      }
    },
...
  },
  "info": {
    "title": "Task Manager API",
    "description": "NestJS Task Manager API Documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "JWT access token"
      }
    },
    "schemas": {
      "ProblemDetailsDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "about:blank",
            "description": "URI reference identifying the problem type. Defaults to about:blank."
          },
          "title": {
            "type": "string",
            "example": "Bad Request",
            "description": "Short, human-readable summary of the problem."
          },
          "status": {
            "type": "number",
            "example": 400,
            "description": "HTTP status code generated by the origin server."
          },
          "detail": {
            "type": "string",
            "example": "Validation failed",
            "description": "Human-readable explanation specific to this occurrence."
          },
          "instance": {
            "type": "string",
            "example": "/api/v1/tasks",
            "description": "URI reference identifying the specific occurrence."
          },
          "requestId": {
            "type": "string",
            "example": "01JXYZ123456789",
            "description": "Optional correlation/request identifier."
          },
          "errors": {
            "example": [
              "email must be an email",
              "password must be longer than or equal to 8 characters"
            ],
            "description": "Optional validation or additional error details.",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance"
        ]
      },
...
    }
  }
}