# src/common/http/problem-details.dto.ts
import { ApiProperty, ApiPropertyOptional } from '@nestjs/swagger';

export class ProblemDetailsDto {
  @ApiProperty({
    example: 'about:blank',
    description:
      'URI reference identifying the problem type. Defaults to about:blank.',
  })
  type!: string;

  @ApiProperty({
    example: 'Bad Request',
    description: 'Short, human-readable summary of the problem.',
  })
  title!: string;

  @ApiProperty({
    example: 400,
    description: 'HTTP status code generated by the origin server.',
  })
  status!: number;

  @ApiProperty({
    example: 'Validation failed',
    description: 'Human-readable explanation specific to this occurrence.',
  })
  detail!: string;

  @ApiProperty({
    example: '/api/v1/tasks',
    description: 'URI reference identifying the specific occurrence.',
  })
  instance!: string;

  @ApiPropertyOptional({
    example: '01JXYZ123456789',
    description: 'Optional correlation/request identifier.',
  })
  requestId?: string;

  @ApiPropertyOptional({
    example: [
      'email must be an email',
      'password must be longer than or equal to 8 characters',
    ],
    description: 'Optional validation or additional error details.',
    type: [String],
  })
  errors?: string[];
}